requirements.txt
python-dotenv>=1.0.0
llama-index>=0.10.0
llama-index-core >= 0.10.0
openai>=1.12.0
pypdf>=4.0.0